% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_open_wbo.R
\name{run_open_wbo}
\alias{run_open_wbo}
\title{Run open-wbo_static}
\usage{
run_open_wbo(args = character())
}
\arguments{
\item{args}{Character vector of arguments passed to `open-wbo_static`.}
}
\value{
Character string containing the output from `open-wbo_static`.
}
\description{
Run the bundled `open-wbo_static` binary with user-supplied parameters.
}
